Added server communication code and methods to new class processclient. Error present regarding not finding data
